+++ lib/logstash/outputs/azure_loganalytics.rb
@@ -1,4 +1,5 @@
 # encoding: utf-8
+
 require "logstash/outputs/base"
 require "logstash/namespace"
 require "stud/buffer"
@@ -14,15 +15,31 @@ class LogStash::Outputs::AzureLogAnalytics < LogStash::Outputs::Base
   # The primary or the secondary Connected Sources client authentication key
   config :shared_key, :validate => :string, :required => true
 
-  # The name of the event type that is being submitted to Log Analytics. This must be only alpha characters.
+  # The name of the event type that is being submitted to Log Analytics. 
+  # This must be only alpha characters.
   config :log_type, :validate => :string, :required => true
 
-  # The name of the time generated field. Be carefule that the value of field should strictly follow the ISO 8601 format (YYYY-MM-DDThh:mm:ssZ)
+  # The service endpoint (Default: ods.opinsights.azure.com)
+  config :endpoint, :validate => :string, :default => 'ods.opinsights.azure.com'
+
+  # The name of the time generated field.
+  # Be carefule that the value of field should strictly follow the ISO 8601 format (YYYY-MM-DDThh:mm:ssZ)
   config :time_generated_field, :validate => :string, :default => ''
 
-  # list of Key names in in-coming record to deliver.
+  # The list of key names in in-coming record that you want to submit to Log Analytics
   config :key_names, :validate => :array, :default => []
-  
+
+  # The list of data types for each column as which you want to store in Log Analytics (`string`, `boolean`, or `double`)
+  # - The key names in `key_types` param must be included in `key_names` param. The column data whose key isn't included in  `key_names` is treated as `string` data type.
+  # - Multiple key value entries are separated by `spaces` rather than commas 
+  #   See also https://www.elastic.co/guide/en/logstash/current/configuration-file-structure.html#hash
+  # - If you want to store a column as datetime or guid data format, set `string` for the column ( the value of the column should be `YYYY-MM-DDThh:mm:ssZ format` if it's `datetime`, and `GUID format` if it's `guid`).
+  # - In case that `key_types` param are not specified, all columns that you want to submit ( you choose with `key_names` param ) are stored as `string` data type in Log Analytics.
+  # Example:
+  #   key_names => ['key1','key2','key3','key4',...]
+  #   key_types => {'key1'=>'string' 'key2'=>'string' 'key3'=>'boolean' 'key4'=>'double' ...}
+  config :key_types, :validate => :hash, :default => {}
+
   # Max number of items to buffer before flushing. Default 50.
   config :flush_items, :validate => :number, :default => 50
   
@@ -38,8 +55,15 @@ class LogStash::Outputs::AzureLogAnalytics < LogStash::Outputs::Base
       raise ArgumentError, 'log_type must be only alpha characters' 
     end
 
+    @key_types.each { |k, v|
+      t = v.downcase
+      if ( !t.eql?('string') && !t.eql?('double') && !t.eql?('boolean') ) 
+        raise ArgumentError, "Key type(#{v}) for key(#{k}) must be either string, boolean, or double"
+      end
+    }
+
     ## Start 
-    @client=Azure::Loganalytics::Datacollectorapi::Client::new(@customer_id,@shared_key)
+    @client=Azure::Loganalytics::Datacollectorapi::Client::new(@customer_id,@shared_key,@endpoint)
 
     buffer_initialize(
       :max_items => @flush_items,
@@ -66,7 +90,11 @@ class LogStash::Outputs::AzureLogAnalytics < LogStash::Outputs::Base
       if @key_names.length > 0
         @key_names.each do |key|
           if event_hash.include?(key)
-            document[key] = event_hash[key]
+            if @key_types.include?(key)
+              document[key] = convert_value(@key_types[key], event_hash[key])
+            else
+              document[key] = event_hash[key]
+            end
           end
         end
       else
@@ -95,7 +123,20 @@ class LogStash::Outputs::AzureLogAnalytics < LogStash::Outputs::Base
     rescue Exception => ex
       @logger.error("Exception occured in posting to DataCollector API: '#{ex}', data=>" + (documents.to_json).to_s)
     end
-
   end # def flush
+
+  private
+  def convert_value(type, val)
+    t = type.downcase
+    case t
+    when "boolean"
+      v = val.downcase
+      return (v.to_s == 'true' ) ? true : false
+    when "double"
+      return Integer(val) rescue Float(val) rescue val
+    else
+      return val
+    end
+  end
 
 end # class LogStash::Outputs::AzureLogAnalytics